Move get_used_cells_and_allocated_sizes to mod
